refactor(notifications): move display-related items in view file
